회원가입 정보 db에 보내는 함수 서버에 추가
