+++ sunghyun/.idea/workspace.xml
@@ -7,7 +7,10 @@
     <list default="true" id="9edecb95-6ea1-400a-ad1c-b3bad67596b0" name="Default Changelist" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/client/src/SignUpPage.js" beforeDir="false" afterPath="$PROJECT_DIR$/client/src/SignUpPage.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/client/src/components/InputInfo.js" beforeDir="false" afterPath="$PROJECT_DIR$/client/src/components/InputInfo.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/client/src/components/SignUpInput.js" beforeDir="false" afterPath="$PROJECT_DIR$/client/src/components/SignUpInput.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/server/index.js" beforeDir="false" afterPath="$PROJECT_DIR$/server/index.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/server/package.json" beforeDir="false" afterPath="$PROJECT_DIR$/server/package.json" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/server/yarn.lock" beforeDir="false" afterPath="$PROJECT_DIR$/server/yarn.lock" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -80,7 +83,7 @@
       <workItem from="1641113967305" duration="599000" />
       <workItem from="1641125845070" duration="7023000" />
       <workItem from="1641280404085" duration="1311000" />
-      <workItem from="1641301351483" duration="2120000" />
+      <workItem from="1641301351483" duration="2835000" />
     </task>
     <servers />
   </component>
