+++ sunghyun/server/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "axios": "^0.24.0",
     "cors": "^2.8.5",
+    "crypto": "^1.0.1",
     "dotenv": "^10.0.0",
     "express": "^4.17.2",
     "mysql2": "^2.3.3"
